chore: 불필요하게 분리된 컴포넌트 파일 제거 후 MyPage에서 헤더, 유저정보 UI 처리
