+++ src/pages/MyPage.tsx
@@ -1,9 +1,17 @@
-import MyPageHeader from '../components/MyPageHeader';
-import BoardList from '../components/BoardList';
-import MyPageUserInfo from '../components/MyPageUserInfo';
-import { BoardInfo, BoardPinInfo } from '../types';
-
 import styled from 'styled-components';
+import { useState, useEffect } from 'react';
+import { IcSearch, icPlus, icSetting } from '../assets/icons';
+import { BoardInfo, BoardPinInfo } from '../types';
+import { FONT_STYLES } from '../styles/font';
+import { COLOR } from '../styles/color';
+import BoardList from '../components/BoardList';
+
+interface MyPageUserInfo {
+  userId: string;
+  nickname: string;
+  followerCnt: number;
+  followingCnt: number;
+}
 
 export default function MyPage() {
   const boardList: BoardPinInfo[] = [
@@ -33,10 +41,34 @@ export default function MyPage() {
     { id: 1, title: '바닷가', boardList: boardList, savedTime: '방금' },
     { id: 2, title: '제주도', boardList: boardList, savedTime: '방금' },
   ];
+  const [userInfo, setUserInfo] = useState<MyPageUserInfo>();
+
+  useEffect(() => {
+    setUserInfo({ userId: 'cheeze123', nickname: '치즈', followerCnt: 36, followingCnt: 54 });
+  }, []);
+
   return (
     <StWrapper>
-      <MyPageHeader />
-      <MyPageUserInfo />
+      <StHeader>
+        <StSearchBarWrapper>
+          <StSearchInput />
+          <StSearchIcon />
+        </StSearchBarWrapper>
+        <StIcon src={icPlus} />
+        <StIcon src={icSetting} />
+      </StHeader>
+      <StProfileImage src="https://images.unsplash.com/photo-1638913658828-afb88c3d4d11?ixlib=rb-1.2.1&ixid=MnwxMjA3fDF8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1170&q=80" />
+      {userInfo && (
+        <StProfileInfoWrapper>
+          <StUserName>{userInfo.nickname}</StUserName>
+          <StUserId>@{userInfo.userId}</StUserId>
+          <StFollowInfoWrapper>
+            <span>팔로워 {userInfo.followerCnt}명</span>
+            <span>팔로잉 {userInfo.followingCnt}명</span>
+          </StFollowInfoWrapper>
+        </StProfileInfoWrapper>
+      )}
+      <StShareButton>공유</StShareButton>
       <BoardList boardList={boardInfo} />
     </StWrapper>
   );
@@ -45,7 +77,78 @@ export default function MyPage() {
 const StWrapper = styled.div`
   display: flex;
   flex-direction: column;
-  width: 100vw;
+  width: 100%;
   align-items: center;
   margin-top: 97px;
+`;
+const StHeader = styled.header`
+  display: flex;
+  width: 100%;
+  height: 72px;
+  justify-content: space-evenly;
+  align-items: center;
+`;
+const StSearchBarWrapper = styled.div`
+  display: flex;
+  align-items: center;
+  position: relative;
+  width: 70%;
+`;
+const StSearchInput = styled.input`
+  width: 100%;
+  height: 48px;
+  border: 2px solid ${COLOR.GRAY_200};
+  border-radius: 24px;
+  padding: 16px 35px;
+`;
+const StSearchIcon = styled(IcSearch)`
+  position: absolute;
+  top: 16px;
+  left: 14px;
+`;
+const StIcon = styled.img`
+  cursor: pointer;
+`;
+const FlexColumnBox = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+`;
+const StProfileImage = styled.img`
+  width: 120px;
+  height: 120px;
+  border-radius: 50%;
+  object-fit: fill;
+  padding-bottom: 3px;
+`;
+const StProfileInfoWrapper = styled(FlexColumnBox)`
+  padding: 3px 0 18px 0;
+`;
+const StUserName = styled.h1`
+  ${FONT_STYLES.H1_BOLD};
+  font-size: 36px;
+`;
+const StUserId = styled.p`
+  ${FONT_STYLES.B2_REGULAR};
+  color: ${COLOR.GRAY_100};
+  font-size: 16px;
+  padding: 2px 0 19px 0;
+`;
+const StFollowInfoWrapper = styled.div`
+  ${FONT_STYLES.B1_BOLD};
+  display: flex;
+  gap: 17px;
+  font-size: 16px;
+`;
+const StShareButton = styled.button`
+  ${FONT_STYLES.B1_BOLD};
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 60px;
+  height: 48px;
+  font-size: 16px;
+  background-color: ${COLOR.GRAY_300};
+  border-radius: 24px;
+  margin-bottom: 52px;
 `;
